% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_application.R
\name{call_shiny}
\alias{call_shiny}
\title{Shiny app function.}
\usage{
call_shiny()
}
\value{
Shiny app with the happiness data analysis and plots.
}
\description{
This function runs the shiny app to display the analysis.
}
\examples{
call_shiny()
}
