% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get_data}
\alias{get_data}
\title{Import Happiness report 2019 Data}
\usage{
get_data()
}
\description{
This function reads the data from a csv-file and returns it as a DataFrame-object.
}
\examples{
To store the data in a variable
df = get_data()
}
