% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_happiness.R
\name{map_happiness}
\alias{map_happiness}
\title{Visualisation of country with its corresponding happiness score.}
\usage{
map_happiness()
}
\description{
This function will create a color coded world map of happiness score,with red being the happiest and purple the least happy countries, using the rworldmap package.
}
